Cleanup: removed IEEP, removed redundant casts, fixed imports, fixed typos.

+++ src/main/java/net/minecraftforge/fml/repackage/com/nothome/delta/GDiffWriter.java
@@ -170,7 +170,7 @@ public class GDiffWriter implements DiffWriter {
     @Override
     public void close() throws IOException {
         this.flush();
-        output.write((byte)EOF);
+        output.write(EOF);
         output.close();
     }
